login_form/requirements.txt
reflex==0.4.0
supabase
python-dotenv
